+++ src/main/kotlin/io/kanro/idea/plugin/protobuf/lang/actions/ProtobufActionGroup.kt
@@ -1,0 +1,9 @@
+package io.kanro.idea.plugin.protobuf.lang.actions
+
+import com.intellij.openapi.actionSystem.DefaultActionGroup
+
+class ProtobufActionGroup : DefaultActionGroup() {
+    override fun hideIfNoVisibleChildren(): Boolean {
+        return true
+    }
+}